Create generalised preventWorkerQueues function
Move createLayoutWorker to worker.js and update it so that it can be used in different scenarious
