+++ src/utils/worker.js
@@ -27,3 +27,30 @@ const createMockWorker = worker => () => {
 };
 
 export const graph = isTest ? createMockWorker(graphWorker) : graphWorker;
+
+/**
+ * Manage the worker, avoiding race conditions by terminating running
+ * processes when a new request is made, and reinitialising the instance.
+ * Example getJob: (instance, payload) => instance.job(payload)
+ * @param {function} worker Init worker and return job functions
+ * @param {function} getJob Callback to select correct job function
+ * @return {function} Function which returns a promise
+ */
+export function preventWorkerQueues(worker, getJob) {
+  let instance = worker();
+  let running = false;
+
+  return payload => {
+    if (running) {
+      // If worker is already processing a job, cancel it and restart
+      instance.terminate();
+      instance = worker();
+    }
+    running = true;
+
+    return getJob(instance, payload).then(response => {
+      running = false;
+      return response;
+    });
+  };
+}
